fix: setting project id from dashboard
